initial jsonrpc client setup
