+++ mint-client/src/clients/user.rs
@@ -16,16 +16,12 @@ use minimint::transaction as mint_tx;
 use minimint::transaction::{Output, TransactionItem};
 use minimint_api::db::batch::DbBatch;
 use minimint_api::db::Database;
-use minimint_api::encoding::Decodable;
 use minimint_api::{Amount, TransactionId};
 use minimint_api::{OutPoint, PeerId};
 use rand::{CryptoRng, RngCore};
 use secp256k1_zkp::{All, Secp256k1};
-use serde::{Deserialize, Serialize};
-use std::collections::VecDeque;
-use std::error::Error;
 use std::sync::{Arc, Mutex};
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::time::Duration;
 use thiserror::Error;
 
 const TIMELOCK: u64 = 100;
@@ -461,197 +457,3 @@ impl From<LnClientError> for ClientError {
         ClientError::LnClientError(e)
     }
 }
-
-// -> clientd
-/// Holds all possible Responses of the RPC-CLient can also be used to parse responses (for client-cli)
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum APIResponse {
-    ///The clients holdings : The quantity of coins for each tier. For total holdings sum(Infoi.quantity * Infoi.tier) with i = 0 - n
-    /// Also contains the [`PendingRes`] variant.
-    Info {
-        coins: Vec<CoinsByTier>,
-        pending: PendingRes,
-    },
-    Pending {
-        pending: PendingRes,
-    },
-    ///Holds a new address for the client to use for peg-in
-    PegInAddress {
-        pegin_address: bitcoin::Address,
-    },
-    ///Holds a [`minimint_api::TransactionId`] from a successful PegIn or PegOut
-    PegIO {
-        txid: TransactionId,
-    },
-    /// Holds the serialized [`Coins<SpendableCoin>`]
-    Spend {
-        token: Coins<SpendableCoin>,
-    },
-    /// Holds the from the federation returned [`OutPoint`] (regarding the reissuance) and the [`TransactionStatus`]
-    Reissue {
-        out_point: OutPoint,
-        status: TransactionStatus,
-    },
-    /// Holds events which could not be sent to the client but were triggered by some action from him. This will be cleared after querying it
-    Events {
-        events: Vec<Event>,
-    },
-    /// Represents an empty response
-    Empty,
-}
-
-/// Active issuances : Not yet (bey the federation) signed BUT accepted coins
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct PendingRes {
-    //TODO: Also return Vec<TransactionId> (?)
-    transactions: usize,
-    acc_qty_coins: usize,
-    acc_val_amount: Amount,
-}
-
-impl PendingRes {
-    pub fn build_pending(all_pending: Vec<CoinFinalizationData>) -> Self {
-        let acc_qty_coins = all_pending.iter().map(|cfd| cfd.coin_count()).sum();
-        let acc_val_amount = all_pending.iter().map(|cfd| cfd.coin_amount()).sum();
-        PendingRes {
-            transactions: all_pending.len(),
-            acc_qty_coins,
-            acc_val_amount,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct Event {
-    timestamp: u64,
-    data: String, //use something else than string ?
-}
-
-impl Event {
-    pub fn build_event(data: String) -> Self {
-        let time = SystemTime::now();
-        let d = time.duration_since(UNIX_EPOCH).unwrap();
-        let timestamp = (d.as_secs() as u64) * 1000 + (u64::from(d.subsec_nanos()) / 1_000_000);
-        Event { timestamp, data }
-    }
-}
-/// Stores events (the (un)successful result of some action initiated by a client which couldn't be sent back to him)
-/// If the capacity is reached events will be dropped from the back
-pub struct EventLog {
-    data: Mutex<VecDeque<Event>>,
-    capacity: usize,
-    //maturity: u64 could be used to drop events with timestamp > maturity <- idea
-}
-
-impl EventLog {
-    pub fn new(capacity: usize) -> Self {
-        EventLog {
-            data: Mutex::new(VecDeque::with_capacity(capacity)),
-            capacity,
-        }
-    }
-    pub fn add(&self, data: String) {
-        let mut events = self.data.lock().unwrap(); // don't know what to do here.. clientd should be restarted if this happens
-                                                    //Because Mutex only guarantees that only one thread at a time but not the (in order) correct one is pushing events
-                                                    //this guarantees that the timestamps will be sorted
-        let event = Event::build_event(data);
-
-        if let Some(ts) = events.back() {
-            if event.timestamp < ts.timestamp {
-                let len = events.len();
-                events.insert(len - 1, event)
-            } else {
-                events.push_back(event);
-            }
-        } else {
-            events.push_back(event);
-        }
-        //If the DeQueue gets too long drop the 'oldest' event
-        if events.len() > self.capacity {
-            events.pop_front();
-            events.shrink_to_fit();
-        }
-    }
-    pub fn get(&self, timestamp: u64) -> Vec<Event> {
-        let events = self.data.lock().unwrap();
-        events
-            .iter()
-            .filter(|e| e.timestamp >= timestamp)
-            .cloned()
-            .collect()
-    }
-}
-
-/// Holds quantity of coins per tier
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct CoinsByTier {
-    tier: u64,
-    quantity: usize,
-}
-/// To Deserialize a peg-in request
-#[derive(Deserialize, Clone, Debug)]
-#[serde(from = "PegInReqRaw")]
-pub struct PegInReq {
-    pub txout_proof: TxOutProof,
-    pub transaction: Transaction,
-}
-#[derive(Deserialize, Clone, Debug)]
-pub struct PegInReqRaw {
-    pub txout_proof: String,
-    pub transaction: String,
-}
-/// To Deserialize a peg-out request (amount in sat)
-#[derive(Deserialize, Clone, Debug)]
-pub struct PegOutReq {
-    pub address: bitcoin::Address,
-    #[serde(with = "bitcoin::util::amount::serde::as_sat")]
-    pub amount: bitcoin::Amount,
-}
-#[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct InvoiceReq {
-    #[serde(with = "crate::ln::serde_invoice")]
-    pub bolt11: lightning_invoice::Invoice,
-}
-impl From<PegInReqRaw> for PegInReq {
-    fn from(raw: PegInReqRaw) -> Self {
-        PegInReq {
-            txout_proof: from_hex(raw.txout_proof.as_str()).unwrap(),
-            transaction: from_hex(raw.transaction.as_str()).unwrap(),
-        }
-    }
-}
-
-fn from_hex<D: Decodable>(s: &str) -> Result<D, Box<dyn Error>> {
-    let bytes = hex::decode(s)?;
-    Ok(D::consensus_decode(std::io::Cursor::new(bytes))?)
-}
-impl APIResponse {
-    /// Builds the [`APIResponse::Info`] variant.
-    pub fn build_info(coins: Coins<SpendableCoin>, cfd: Vec<CoinFinalizationData>) -> Self {
-        let info_coins: Vec<CoinsByTier> = coins
-            .coins
-            .iter()
-            .map(|(tier, c)| CoinsByTier {
-                quantity: c.len(),
-                tier: tier.milli_sat,
-            })
-            .collect();
-        APIResponse::Info {
-            coins: info_coins,
-            pending: PendingRes::build_pending(cfd),
-        }
-    }
-    /// Builds the [`APIResponse::Spend`] variant.
-    pub fn build_spend(token: Coins<SpendableCoin>) -> Self {
-        APIResponse::Spend { token }
-    }
-    /// Builds the [`APIResponse::Reissue`] variant.
-    pub fn build_reissue(out_point: OutPoint, status: TransactionStatus) -> Self {
-        APIResponse::Reissue { out_point, status }
-    }
-    /// Builds the [`APIResponse::Events`] variant.
-    pub fn build_events(events: Vec<Event>) -> Self {
-        APIResponse::Events { events }
-    }
-}
-// <- clientd
