--- mint-client/src/rpc/jsonrpc.rs
@@ -1,1 +1,0 @@
-
